requirements.txt
google-generativeai
langgraph
psycopg2-binary
